Debugger optimized for speedier breakpoint processing. Added debugger compile constant to enable or disable it if speed is an issue. Standard gamepads now return 1s when their data register is used up like authentic controllers. Fourscore signature bits tweaked. Fourscore bug was prevent it from ever being enabled. Various performance tweaks focused at PPU background rendering. Palette is now applied within the PPU and frameBuffer bitmap is gone. Zapper light detection method changed due to screen buffer changes, will now be affected by palette file picked. colorChart is now an integer array instead of Color, ToARBG() was hard on performance. NameTable and PatternTable viewers optimized. Keyboard input handler optimized. ScreenCRC caluclator is now unreliable.

+++ DirectXEmu/AddBreakpoint.cs
@@ -12,7 +12,7 @@ namespace DirectXEmu
     public partial class AddBreakpoint : Form
     {
         public int address;
-        public int type;
+        public byte type;
         public AddBreakpoint()
         {
             InitializeComponent();
